Test start/stop logic, rename restart to reset

+++ greenmail-core/src/main/java/com/icegreen/greenmail/util/Service.java
@@ -70,7 +70,7 @@ abstract public class Service extends Thread {
                 if (null == millis) {
                     join();
                 } else {
-                    join(millis.longValue());
+                    join(millis);
                 }
             }
         } catch (InterruptedException e) {
